requirements.txt
streamlit>=1.32
streamlit-authenticator==0.3.2
python-dotenv>=1.0.0
Pillow>=10.0.0
PyYAML>=6.0
snowflake-connector-python>=3.0.0

